backend/backend/fresh/requirements.txt
Django==5.2.4
djangorestframework==3.14.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-decouple==3.8
Pillow==11.3.0 